test(nuxt): fold clearNuxtData coverage into use-async-data tests
